Module rapier3d::control

Utilities for controlling the trajectories of objects in a non-physical way.

Structs

CharacterAutostep

Configuration for the auto-stepping character controller feature.

CharacterCollision

A collision between the character and its environment during its movement.

DynamicRayCastVehicleController

A character controller to simulate vehicles using ray-casting for the wheels.

EffectiveCharacterMovement

The effective movement computed by the character controller.

KinematicCharacterController

A character controller for kinematic bodies.

Wheel

A wheel attached to a vehicle.

WheelTuning

Parameters affecting the physical behavior of a wheel.

Enums

CharacterLength

A length measure used for various options of a character controller.