Struct rapier3d::counters::Counters

pub struct Counters {
    pub enabled: bool,
    pub step_time: Timer,
    pub custom: Timer,
    pub stages: StagesCounters,
    pub cd: CollisionDetectionCounters,
    pub solver: SolverCounters,
    pub ccd: CCDCounters,
}

Aggregation of all the performances counters tracked by rapier.

Fields
enabled: bool

Whether thi counter is enabled or not.

step_time: Timer

Timer for a whole timestep.

custom: Timer

Timer used for debugging.

stages: StagesCounters

Counters of every satge of one time step.

cd: CollisionDetectionCounters

Counters of the collision-detection stage.

solver: SolverCounters

Counters of the constraints resolution and force computation stage.

ccd: CCDCounters

Counters for the CCD resolution stage.

Implementations

impl Counters

pub fn new(enabled: bool) -> Self

Create a new set of counters initialized to zero.

pub fn enable(&mut self)

Enable all the counters.

pub fn enabled(&self) -> bool

Return true if the counters are enabled.

pub fn disable(&mut self)

Disable all the counters.

pub fn step_started(&mut self)

Notify that the time-step has started.

pub fn step_completed(&mut self)

Notfy that the time-step has finished.

pub fn step_time(&self) -> f64

Total time spent for one of the physics engine.

pub fn custom_started(&mut self)

Notify that the custom operation has started.

pub fn custom_completed(&mut self)

Notfy that the custom operation has finished.

pub fn custom_time(&self) -> f64

Total time of a custom event.

pub fn set_nconstraints(&mut self, n: usize)

Set the number of constraints generated.

pub fn set_ncontacts(&mut self, n: usize)

Set the number of contacts generated.

pub fn set_ncontact_pairs(&mut self, n: usize)

Set the number of contact pairs generated.

pub fn reset(&mut self)

Resets all the counters and timers.

impl Counters

pub fn update_started(&mut self)

Start this timer.

pub fn update_completed(&mut self)

Stop this timer.

pub fn update_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn collision_detection_started(&mut self)

Start this timer.

pub fn collision_detection_completed(&mut self)

Stop this timer.

pub fn collision_detection_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn island_construction_started(&mut self)

Start this timer.

pub fn island_construction_completed(&mut self)

Stop this timer.

pub fn island_construction_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn solver_started(&mut self)

Start this timer.

pub fn solver_completed(&mut self)

Stop this timer.

pub fn solver_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn ccd_started(&mut self)

Start this timer.

pub fn ccd_completed(&mut self)

Stop this timer.

pub fn ccd_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn assembly_started(&mut self)

Start this timer.

pub fn assembly_completed(&mut self)

Stop this timer.

pub fn assembly_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn velocity_resolution_started(&mut self)

Start this timer.

pub fn velocity_resolution_completed(&mut self)

Stop this timer.

pub fn velocity_resolution_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn velocity_update_started(&mut self)

Start this timer.

pub fn velocity_update_completed(&mut self)

Stop this timer.

pub fn velocity_update_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn position_resolution_started(&mut self)

Start this timer.

pub fn position_resolution_completed(&mut self)

Stop this timer.

pub fn position_resolution_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn broad_phase_started(&mut self)

Start this timer.

pub fn broad_phase_completed(&mut self)

Stop this timer.

pub fn broad_phase_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn narrow_phase_started(&mut self)

Start this timer.

pub fn narrow_phase_completed(&mut self)

Stop this timer.

pub fn narrow_phase_time(&self) -> f64

Gets the time elapsed for this timer.