Struct rapier3d::dynamics::RopeJoint

#[repr(transparent)]
pub struct RopeJoint {
    pub data: GenericJoint,
}

A rope joint, limits the maximum distance between two bodies

Fields
data: GenericJoint

The underlying joint data.

Implementations

impl RopeJoint

pub fn new() -> Self

Creates a new rope joint limiting the max distance between to bodies

pub fn data(&self) -> &GenericJoint

The underlying generic joint.

pub fn contacts_enabled(&self) -> bool

Are contacts between the attached rigid-bodies enabled?

pub fn set_contacts_enabled(&mut self, enabled: bool) -> &mut Self

Sets whether contacts between the attached rigid-bodies are enabled.

pub fn local_anchor1(&self) -> Point<Real>

The joint’s anchor, expressed in the local-space of the first rigid-body.

pub fn set_local_anchor1(&mut self, anchor1: Point<Real>) -> &mut Self

Sets the joint’s anchor, expressed in the local-space of the first rigid-body.

pub fn local_anchor2(&self) -> Point<Real>

The joint’s anchor, expressed in the local-space of the second rigid-body.

pub fn set_local_anchor2(&mut self, anchor2: Point<Real>) -> &mut Self

Sets the joint’s anchor, expressed in the local-space of the second rigid-body.

pub fn local_axis1(&self) -> UnitVector<Real>

The principal axis of the joint, expressed in the local-space of the first rigid-body.

pub fn set_local_axis1(&mut self, axis1: UnitVector<Real>) -> &mut Self

Sets the principal axis of the joint, expressed in the local-space of the first rigid-body.

pub fn local_axis2(&self) -> UnitVector<Real>

The principal axis of the joint, expressed in the local-space of the second rigid-body.

pub fn set_local_axis2(&mut self, axis2: UnitVector<Real>) -> &mut Self

Sets the principal axis of the joint, expressed in the local-space of the second rigid-body.

pub fn motor(&self, axis: JointAxis) -> Option<&JointMotor>

The motor affecting the joint’s translational degree of freedom.

pub fn set_motor_model(&mut self, model: MotorModel) -> &mut Self

Set the spring-like model used by the motor to reach the desired target velocity and position.

pub fn set_motor_velocity( &mut self, target_vel: Real, factor: Real ) -> &mut Self

Sets the target velocity this motor needs to reach.

pub fn set_motor_position( &mut self, target_pos: Real, stiffness: Real, damping: Real ) -> &mut Self

Sets the target angle this motor needs to reach.

pub fn set_motor( &mut self, target_pos: Real, target_vel: Real, stiffness: Real, damping: Real ) -> &mut Self

Configure both the target angle and target velocity of the motor.

pub fn set_motor_max_force(&mut self, max_force: Real) -> &mut Self

Sets the maximum force the motor can deliver.

pub fn limits(&self, axis: JointAxis) -> Option<&JointLimits<Real>>

The limit maximum distance attached bodies can translate.

pub fn set_limits(&mut self, limits: [Real; 2]) -> &mut Self

Sets the [min,max] limit distances attached bodies can translate.