Type Alias rapier3d::dynamics::BodyStatus

pub type BodyStatus = RigidBodyType;

👎Deprecated: renamed as RigidBodyType

The type of a body, governing the way it is affected by external forces.

Aliased Type

enum BodyStatus {
    Dynamic = 0,
    Fixed = 1,
    KinematicPositionBased = 2,
    KinematicVelocityBased = 3,
}

Variants

Dynamic = 0

A RigidBodyType::Dynamic body can be affected by all external forces.

Fixed = 1

A RigidBodyType::Fixed body cannot be affected by external forces.

KinematicPositionBased = 2

A RigidBodyType::KinematicPositionBased body cannot be affected by any external forces but can be controlled by the user at the position level while keeping realistic one-way interaction with dynamic bodies.

One-way interaction means that a kinematic body can push a dynamic body, but a kinematic body cannot be pushed by anything. In other words, the trajectory of a kinematic body can only be modified by the user and is independent from any contact or joint it is involved in.

KinematicVelocityBased = 3

A RigidBodyType::KinematicVelocityBased body cannot be affected by any external forces but can be controlled by the user at the velocity level while keeping realistic one-way interaction with dynamic bodies.

One-way interaction means that a kinematic body can push a dynamic body, but a kinematic body cannot be pushed by anything. In other words, the trajectory of a kinematic body can only be modified by the user and is independent from any contact or joint it is involved in.