Module rapier3d::geometry

Structures related to geometry: colliders, shapes, etc.

Structs

ActiveCollisionTypes

Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

Ball

A Ball shape.

BroadPhase

A broad-phase combining a Hierarchical Grid and Sweep-and-Prune.

Capsule

A capsule shape defined as a round segment.

Collider

A geometric entity that can be attached to a body so it can be affected by contacts and proximity queries.

ColliderBroadPhaseData

Data associated to a collider that takes part to a broad-phase algorithm.

ColliderBuilder

A structure responsible for building a new collider.

ColliderChanges

Flags describing how the collider has been modified by the user.

ColliderFlags

A set of flags for controlling collision/intersection filtering, modification, and events.

ColliderHandle

The unique identifier of a collider added to a collider set.

ColliderMaterial

The constraints solver-related properties of this collider (friction, restitution, etc.)

ColliderPair

A pair of collider handles.

ColliderParent

Information about the rigid-body this collider is attached to.

ColliderPosition

The position of a collider.

ColliderSet

A set of colliders that can be handled by a physics World.

CollisionEventFlags

Flags providing more information regarding a collision event.

Compound

A compound shape with an aabb bounding volume.

Cone

Cone shape with its principal axis aligned with the y axis.

ContactData

A single contact between two collider.

ContactForceEvent

Event occurring when the sum of the magnitudes of the contact forces between two colliders exceed a threshold.

ContactManifoldData

A contact manifold between two colliders.

ContactPair

The description of all the contacts between a pair of colliders.

ConvexPolyhedron

A convex polyhedron without degenerate faces.

Cuboid

Shape of a box.

Cylinder

Cylinder shape with its principal axis aligned with the y axis.

GenericHeightField

A 3D heightfield with a generic storage buffer for its height grid.

HalfSpace

A half-space delimited by an infinite plane.

HeightFieldCellStatus

The status of the cell of an heightfield.

HeightFieldRadialTriangles

An iterator through all the triangles around the given point, after vertical projection on the heightfield.

InteractionGraph

A graph where nodes are collision objects and edges are contact or proximity algorithms.

InteractionGroups

Pairwise filtering using bit masks.

IntersectionPair

The description of all the contacts between a pair of colliders.

NarrowPhase

The narrow-phase responsible for computing precise contact information between colliders.

PolygonalFeature

A polygonal feature representing the local polygonal approximation of a vertex, face, or edge of a convex shape.

Polyline

A polyline.

RoundShape

A shape with rounded borders.

Segment

A segment shape.

SharedShape

The shape of a collider.

SolverContact

A contact seen by the constraints solver for computing forces.

SolverFlags

Flags affecting the behavior of the constraints solver for a given contact manifold.

Tetrahedron

A tetrahedron with 4 vertices.

TopoFace

A face of a triangle-mesh’s half-edge topology.

TopoHalfEdge

A half-edge of a triangle-mesh’s half-edge topology.

TopoVertex

A vertex of a triangle-mesh’s half-edge topology.

TrackedContact

A single contact between two shape.

TriMesh

A triangle mesh.

TriMeshConnectedComponents

The connected-components of a triangle mesh.

TriMeshFlags

The status of the cell of an heightfield.

TriMeshPseudoNormals

The set of pseudo-normals of a triangle mesh.

TriMeshTopology

The half-edge topology information of a triangle mesh.

Triangle

A triangle shape.

Enums

BroadPhasePairEvent

An event emitted by the broad-phase.

ColliderMassProps

The mass-properties of a collider.

ColliderType

The type of collider.

CollisionEvent

Events occurring when two colliders start or stop colliding

FeatureId

An identifier of a feature of a convex polyhedron.

SegmentPointLocation

Logical description of the location of a point on a triangle.

ShapeType

Enum representing the type of a shape.

TetrahedronPointLocation

Logical description of the location of a point on a triangle.

TopologyError

Indicated an inconsistency in the topology of a triangle mesh.

TriangleOrientation

Orientation of a triangle.

TrianglePointLocation

Description of the location of a point on a triangle.

TypedShape

Enum representing the shape with its actual type

Traits

HeightFieldStorage

Abstraction over the storage type of an heightfield’s heights grid.

PolygonalFeatureMap

Trait implemented by convex shapes with features with polyhedral approximations.

Shape

Trait implemented by shapes usable by Rapier.

SimdCompositeShape

Trait implemented by shapes composed of multiple simpler shapes.

SupportMap

Traits of convex shapes representable by a support mapping function.

TypedSimdCompositeShape

Type Definitions

AABB

An axis-aligned bounding box.

Ball

A ball shape.

Capsule

A capsule shape.

ColliderGraphIndex

Index of a node of the interaction graph.

ColliderShape

The shape of a collider.

Cone

A cone shape.

Contact

A contact between two colliders.

ContactManifold

A contact manifold between two colliders.

Cuboid

A cuboid shape.

Cylinder

A cylindrical shape.

HeightField

A heightfield shape.

PointProjection

The the projection of a point on a collider.

Ray

A ray that can be cast against colliders.

RayIntersection

The intersection between a ray and a collider.

RigidBodyGraphIndex

Index of a node of the interaction graph.

RoundCone

A cone dilated by a sphere (so it has round corners).

RoundConvexPolyhedron

A convex polyhedron dilated by a sphere (so it has round corners).

RoundCuboid

A cuboid dilated by a sphere (so it has round corners).

RoundCylinder

A cylinder dilated by a sphere (so it has round corners).

RoundTriangle

A triangle dilated by a sphere (so it has round corners).

Segment

A segment shape.

TOI

The the time of impact between two shapes.

TemporaryInteractionIndex

Temporary index to and edge of the interaction graph.

Triangle

A triangle shape.