Struct rapier3d::geometry::Collider

pub struct Collider {
    pub user_data: u128,
    /* private fields */
}

A geometric entity that can be attached to a body so it can be affected by contacts and proximity queries.

To build a new collider, use the ColliderBuilder structure.

Fields
user_data: u128

User-defined data associated to this collider.

Implementations

impl Collider

pub fn parent(&self) -> Option<RigidBodyHandle>

The rigid body this collider is attached to.

pub fn is_sensor(&self) -> bool

Is this collider a sensor?

pub fn active_hooks(&self) -> ActiveHooks

The physics hooks enabled for this collider.

pub fn set_active_hooks(&mut self, active_hooks: ActiveHooks)

Sets the physics hooks enabled for this collider.

pub fn active_events(&self) -> ActiveEvents

The events enabled for this collider.

pub fn set_active_events(&mut self, active_events: ActiveEvents)

Sets the events enabled for this collider.

pub fn active_collision_types(&self) -> ActiveCollisionTypes

The collision types enabled for this collider.

pub fn set_active_collision_types(
&mut self,
active_collision_types: ActiveCollisionTypes
)

Sets the collision types enabled for this collider.

pub fn friction(&self) -> Real

The friction coefficient of this collider.

pub fn set_friction(&mut self, coefficient: Real)

Sets the friction coefficient of this collider.

pub fn friction_combine_rule(&self) -> CoefficientCombineRule

The combine rule used by this collider to combine its friction coefficient with the friction coefficient of the other collider it is in contact with.

pub fn set_friction_combine_rule(&mut self, rule: CoefficientCombineRule)

Sets the combine rule used by this collider to combine its friction coefficient with the friction coefficient of the other collider it is in contact with.

pub fn restitution(&self) -> Real

The restitution coefficient of this collider.

pub fn set_restitution(&mut self, coefficient: Real)

Sets the restitution coefficient of this collider.

pub fn restitution_combine_rule(&self) -> CoefficientCombineRule

The combine rule used by this collider to combine its restitution coefficient with the restitution coefficient of the other collider it is in contact with.

pub fn set_restitution_combine_rule(&mut self, rule: CoefficientCombineRule)

Sets the combine rule used by this collider to combine its restitution coefficient with the restitution coefficient of the other collider it is in contact with.

pub fn set_contact_force_event_threshold(&mut self, threshold: Real)

Sets the total force magnitude beyond which a contact force event can be emitted.

pub fn set_sensor(&mut self, is_sensor: bool)

Sets whether or not this is a sensor collider.

pub fn set_translation(&mut self, translation: Vector<Real>)

Sets the translational part of this collider’s position.

pub fn set_rotation(&mut self, rotation: AngVector<Real>)

Sets the rotational part of this collider’s position.

pub fn set_position(&mut self, position: Isometry<Real>)

Sets the position of this collider.

pub fn position(&self) -> &Isometry<Real>

The world-space position of this collider.

pub fn translation(&self) -> &Vector<Real>

The translational part of this collider’s position.

pub fn rotation(&self) -> &Rotation<Real>

The rotational part of this collider’s position.

pub fn position_wrt_parent(&self) -> Option<&Isometry<Real>>

The position of this collider wrt the body it is attached to.

pub fn set_translation_wrt_parent(&mut self, translation: Vector<Real>)

Sets the translational part of this collider’s translation relative to its parent rigid-body.

pub fn set_rotation_wrt_parent(&mut self, rotation: AngVector<Real>)

Sets the rotational part of this collider’s rotaiton relative to its parent rigid-body.

pub fn set_position_wrt_parent(&mut self, pos_wrt_parent: Isometry<Real>)

Sets the position of this collider wrt. its parent rigid-body.

Does nothing if the collider is not attached to a rigid-body.

pub fn collision_groups(&self) -> InteractionGroups

The collision groups used by this collider.

pub fn set_collision_groups(&mut self, groups: InteractionGroups)

Sets the collision groups of this collider.

pub fn solver_groups(&self) -> InteractionGroups

The solver groups used by this collider.

pub fn set_solver_groups(&mut self, groups: InteractionGroups)

Sets the solver groups of this collider.

pub fn material(&self) -> &ColliderMaterial

The material (friction and restitution properties) of this collider.

pub fn volume(&self) -> Real

The volume (or surface in 2D) of this collider.

pub fn density(&self) -> Real

The density of this collider.

pub fn mass(&self) -> Real

The mass of this collider.

pub fn set_density(&mut self, density: Real)

Sets the uniform density of this collider.

This will override any previous mass-properties set by Self::set_density, Self::set_mass, Self::set_mass_properties, ColliderBuilder::density, ColliderBuilder::mass, or ColliderBuilder::mass_properties for this collider.

The mass and angular inertia of this collider will be computed automatically based on its shape.

pub fn set_mass(&mut self, mass: Real)

Sets the mass of this collider.

This will override any previous mass-properties set by Self::set_density, Self::set_mass, Self::set_mass_properties, ColliderBuilder::density, ColliderBuilder::mass, or ColliderBuilder::mass_properties for this collider.

The angular inertia of this collider will be computed automatically based on its shape and this mass value.

pub fn set_mass_properties(&mut self, mass_properties: MassProperties)

Sets the mass properties of this collider.

This will override any previous mass-properties set by Self::set_density, Self::set_mass, Self::set_mass_properties, ColliderBuilder::density, ColliderBuilder::mass, or ColliderBuilder::mass_properties for this collider.

pub fn shape(&self) -> &dyn Shape

The geometric shape of this collider.

pub fn shape_mut(&mut self) -> &mut dyn Shape

A mutable reference to the geometric shape of this collider.

If that shape is shared by multiple colliders, it will be cloned first so that self contains a unique copy of that shape that you can modify.

pub fn set_shape(&mut self, shape: SharedShape)

Sets the shape of this collider.

pub fn shared_shape(&self) -> &SharedShape

Retrieve the SharedShape. Also see the shape() function

pub fn compute_aabb(&self) -> AABB

Compute the axis-aligned bounding box of this collider.

pub fn compute_swept_aabb(&self, next_position: &Isometry<Real>) -> AABB

Compute the axis-aligned bounding box of this collider moving from its current position to the given next_position

pub fn mass_properties(&self) -> MassProperties

Compute the local-space mass properties of this collider.

pub fn contact_force_event_threshold(&self) -> Real

The total force magnitude beyond which a contact force event can be emitted.