Struct rapier3d::geometry::ColliderBuilder

pub struct ColliderBuilder {

    pub shape: SharedShape,
    pub mass_properties: ColliderMassProps,
    pub friction: Real,
    pub friction_combine_rule: CoefficientCombineRule,
    pub restitution: Real,
    pub restitution_combine_rule: CoefficientCombineRule,
    pub position: Isometry<Real>,
    pub is_sensor: bool,
    pub active_collision_types: ActiveCollisionTypes,
    pub active_hooks: ActiveHooks,
    pub active_events: ActiveEvents,
    pub user_data: u128,
    pub collision_groups: InteractionGroups,
    pub solver_groups: InteractionGroups,
    pub contact_force_event_threshold: Real,

}

A structure responsible for building a new collider.

Fields
shape: SharedShape

The shape of the collider to be built.

mass_properties: ColliderMassProps

Controls the way the collider’s mass-properties are computed.

friction: Real

The friction coefficient of the collider to be built.

friction_combine_rule: CoefficientCombineRule

The rule used to combine two friction coefficients.

restitution: Real

The restitution coefficient of the collider to be built.

restitution_combine_rule: CoefficientCombineRule

The rule used to combine two restitution coefficients.

position: Isometry<Real>

The position of this collider.

is_sensor: bool

Is this collider a sensor?

active_collision_types: ActiveCollisionTypes

Contact pairs enabled for this collider.

active_hooks: ActiveHooks

Physics hooks enabled for this collider.

active_events: ActiveEvents

Events enabled for this collider.

user_data: u128

The user-data of the collider being built.

collision_groups: InteractionGroups

The collision groups for the collider being built.

solver_groups: InteractionGroups

The solver groups for the collider being built.

contact_force_event_threshold: Real

The total force magnitude beyond which a contact force event can be emitted.

Implementations

impl ColliderBuilder

pub fn new(shape: SharedShape) -> Self

Initialize a new collider builder with the given shape.

pub fn compound(shapes: Vec<(Isometry<Real>, SharedShape)>) -> Self

Initialize a new collider builder with a compound shape.

pub fn ball(radius: Real) -> Self

Initialize a new collider builder with a ball shape defined by its radius.

pub fn halfspace(outward_normal: Unit<Vector<Real>>) -> Self

Initialize a new collider build with a half-space shape defined by the outward normal of its planar boundary.

pub fn cylinder(half_height: Real, radius: Real) -> Self

Initialize a new collider builder with a cylindrical shape defined by its half-height (along along the y axis) and its radius.

pub fn round_cylinder(
half_height: Real,
radius: Real,
border_radius: Real
) -> Self

Initialize a new collider builder with a rounded cylindrical shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn cone(half_height: Real, radius: Real) -> Self

Initialize a new collider builder with a cone shape defined by its half-height (along along the y axis) and its basis radius.

pub fn round_cone(half_height: Real, radius: Real, border_radius: Real) -> Self

Initialize a new collider builder with a rounded cone shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn capsule_x(half_height: Real, radius: Real) -> Self

Initialize a new collider builder with a capsule shape aligned with the x axis.

pub fn capsule_y(half_height: Real, radius: Real) -> Self

Initialize a new collider builder with a capsule shape aligned with the y axis.

pub fn capsule_z(half_height: Real, radius: Real) -> Self

Initialize a new collider builder with a capsule shape aligned with the z axis.

pub fn cuboid(hx: Real, hy: Real, hz: Real) -> Self

Initialize a new collider builder with a cuboid shape defined by its half-extents.

pub fn round_cuboid(hx: Real, hy: Real, hz: Real, border_radius: Real) -> Self

Initialize a new collider builder with a round cuboid shape defined by its half-extents and border radius.

pub fn segment(a: Point<Real>, b: Point<Real>) -> Self

Initializes a collider builder with a segment shape.

pub fn triangle(a: Point<Real>, b: Point<Real>, c: Point<Real>) -> Self

Initializes a collider builder with a triangle shape.

pub fn round_triangle(
a: Point<Real>,
b: Point<Real>,
c: Point<Real>,
border_radius: Real
) -> Self

Initializes a collider builder with a triangle shape with round corners.

pub fn polyline(
vertices: Vec<Point<Real>>,
indices: Option<Vec<[u32; 2]>>
) -> Self

Initializes a collider builder with a polyline shape defined by its vertex and index buffers.

pub fn trimesh(vertices: Vec<Point<Real>>, indices: Vec<[u32; 3]>) -> Self

Initializes a collider builder with a triangle mesh shape defined by its vertex and index buffers.

pub fn convex_decomposition(
vertices: &[Point<Real>],
indices: &[[u32; 3]]
) -> Self

Initializes a collider builder with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
border_radius: Real
) -> Self

Initializes a collider builder with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_decomposition_with_params(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
params: &VHACDParameters
) -> Self

Initializes a collider builder with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition_with_params(
vertices: &[Point<Real>],
indices: &[[u32; 3]],
params: &VHACDParameters,
border_radius: Real
) -> Self

Initializes a collider builder with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_hull(points: &[Point<Real>]) -> Option<Self>

Initializes a new collider builder with a 2D convex polygon or 3D convex polyhedron obtained after computing the convex-hull of the given points.

pub fn round_convex_hull(
points: &[Point<Real>],
border_radius: Real
) -> Option<Self>

Initializes a new collider builder with a round 2D convex polygon or 3D convex polyhedron obtained after computing the convex-hull of the given points. The shape is dilated by a sphere of radius border_radius.

pub fn convex_mesh(
points: Vec<Point<Real>>,
indices: &[[u32; 3]]
) -> Option<Self>

Creates a new collider builder that is a convex polyhedron formed by the given triangle-mesh assumed to be convex (no convex-hull will be automatically computed).

pub fn round_convex_mesh(
points: Vec<Point<Real>>,
indices: &[[u32; 3]],
border_radius: Real
) -> Option<Self>

Creates a new collider builder that is a round convex polyhedron formed by the given triangle-mesh assumed to be convex (no convex-hull will be automatically computed). The triangle mesh shape is dilated by a sphere of radius border_radius.

pub fn heightfield(heights: DMatrix<Real>, scale: Vector<Real>) -> Self

Initializes a collider builder with a heightfield shape defined by its set of height and a scale factor along each coordinate axis.

pub fn default_friction() -> Real

The default friction coefficient used by the collider builder.

pub fn default_density() -> Real

The default density used by the collider builder.

pub fn user_data(self, data: u128) -> Self

Sets an arbitrary user-defined 128-bit integer associated to the colliders built by this builder.

pub fn collision_groups(self, groups: InteractionGroups) -> Self

Sets the collision groups used by this collider.

Two colliders will interact iff. their collision groups are compatible. See InteractionGroups::test for details.

pub fn solver_groups(self, groups: InteractionGroups) -> Self

Sets the solver groups used by this collider.

Forces between two colliders in contact will be computed iff their solver groups are compatible. See InteractionGroups::test for details.

pub fn sensor(self, is_sensor: bool) -> Self

Sets whether or not the collider built by this builder is a sensor.

pub fn active_hooks(self, active_hooks: ActiveHooks) -> Self

The set of physics hooks enabled for this collider.

pub fn active_events(self, active_events: ActiveEvents) -> Self

The set of events enabled for this collider.

pub fn active_collision_types(
self,
active_collision_types: ActiveCollisionTypes
) -> Self

The set of active collision types for this collider.

pub fn friction(self, friction: Real) -> Self

Sets the friction coefficient of the collider this builder will build.

pub fn friction_combine_rule(self, rule: CoefficientCombineRule) -> Self

Sets the rule to be used to combine two friction coefficients in a contact.

pub fn restitution(self, restitution: Real) -> Self

Sets the restitution coefficient of the collider this builder will build.

pub fn restitution_combine_rule(self, rule: CoefficientCombineRule) -> Self

Sets the rule to be used to combine two restitution coefficients in a contact.

pub fn density(self, density: Real) -> Self

Sets the uniform density of the collider this builder will build.

This will be overridden by a call to Self::mass or Self::mass_properties so it only makes sense to call either Self::density or Self::mass or Self::mass_properties.

The mass and angular inertia of this collider will be computed automatically based on its shape.

pub fn mass(self, mass: Real) -> Self

Sets the mass of the collider this builder will build.

This will be overridden by a call to Self::density or Self::mass_properties so it only makes sense to call either Self::density or Self::mass or Self::mass_properties.

The angular inertia of this collider will be computed automatically based on its shape and this mass value.

pub fn mass_properties(self, mass_properties: MassProperties) -> Self

Sets the mass properties of the collider this builder will build.

This will be overridden by a call to Self::density or Self::mass so it only makes sense to call either Self::density or Self::mass or Self::mass_properties.

pub fn contact_force_event_threshold(self, threshold: Real) -> Self

Sets the total force magnitude beyond which a contact force event can be emitted.

pub fn translation(self, translation: Vector<Real>) -> Self

Sets the initial translation of the collider to be created.

If the collider will be attached to a rigid-body, this sets the translation relative to the rigid-body it will be attached to.

pub fn rotation(self, angle: AngVector<Real>) -> Self

Sets the initial orientation of the collider to be created.

If the collider will be attached to a rigid-body, this sets the orientation relative to the rigid-body it will be attached to.

pub fn position(self, pos: Isometry<Real>) -> Self

Sets the initial position (translation and orientation) of the collider to be created.

If the collider will be attached to a rigid-body, this sets the position relative to the rigid-body it will be attached to.

pub fn position_wrt_parent(self, pos: Isometry<Real>) -> Self

👎 Deprecated:
Use .position instead.

Sets the initial position (translation and orientation) of the collider to be created, relative to the rigid-body it is attached to.

pub fn delta(self, delta: Isometry<Real>) -> Self

👎 Deprecated:
Use .position instead.

Set the position of this collider in the local-space of the rigid-body it is attached to.

pub fn build(&self) -> Collider

Builds a new collider attached to the given rigid-body.