Struct rapier3d::geometry::Polyline

pub struct Polyline { /* private fields */ }

A polyline.

Implementations

impl Polyline

pub fn bounding_sphere( &self, pos: &Isometry<f32, Unit<Quaternion<f32>>, 3> ) -> BoundingSphere

Computes the world-space bounding sphere of this polyline, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this polyline.

impl Polyline

pub fn new( vertices: Vec<OPoint<f32, Const<3>>, Global>, indices: Option<Vec<[u32; 2], Global>> ) -> Polyline

Creates a new polyline from a vertex buffer and an index buffer.

pub fn aabb(&self, pos: &Isometry<f32, Unit<Quaternion<f32>>, 3>) -> Aabb

Compute the axis-aligned bounding box of this polyline.

pub fn local_aabb(&self) -> &Aabb

Gets the local axis-aligned bounding box of this polyline.

pub fn num_segments(&self) -> usize

The number of segments forming this polyline.

pub fn segments(&self) -> impl ExactSizeIterator

An iterator through all the segments of this mesh.

pub fn segment(&self, i: u32) -> Segment

Get the i-th segment of this mesh.

pub fn segment_feature_to_polyline_feature( &self, segment: u32, _feature: FeatureId ) -> FeatureId

Transforms the feature-id of a segment to the feature-id of this polyline.

pub fn vertices(&self) -> &[OPoint<f32, Const<3>>]

The vertex buffer of this mesh.

pub fn indices(&self) -> &[[u32; 2]]

The index buffer of this mesh.

pub fn flat_indices(&self) -> &[u32]

A flat view of the index buffer of this mesh.

pub fn scaled( self, scale: &Matrix<f32, Const<3>, Const<1>, ArrayStorage<f32, 3, 1>> ) -> Polyline

Computes a scaled version of this polyline.

pub fn reverse(&mut self)

Reverse the orientation of this polyline by swapping the indices of all its segments and reverting its index buffer.

pub fn extract_connected_components(&self) -> Vec<Polyline, Global>

Extracts the connected components of this polyline, consuming self.

This method is currently quite restrictive on the kind of allowed input. The polyline represented by self must already have an index buffer sorted such that:

Each connected component appears in the index buffer one after the other, i.e., a connected component of this polyline must be a contiguous range of this polyline’s index buffer.

Each connected component is closed, i.e., each range of this polyline index buffer self.indices[i_start..=i_end] forming a complete connected component, we must have self.indices[i_start][0] == self.indices[i_end][1].

The indices for each component must already be in order, i.e., if the segments self.indices[i] and self.indices[i + 1] are part of the same connected component then we must have self.indices[i][1] == self.indices[i + 1][0].

Output

Returns the set of polylines. If the inputs fulfill the constraints mentioned above, each polyline will be a closed loop with consistent edge orientations, i.e., for all indices i, we have polyline.indices[i][1] == polyline.indices[i + 1][0].

The orientation of each closed loop (clockwise or counterclockwise) are identical to their original orientation in self.

pub fn project_local_point_assuming_solid_interior_ccw( &self, point: OPoint<f32, Const<3>>, axis: u8 ) -> (PointProjection, (u32, SegmentPointLocation))

Perform a point projection assuming a solid interior based on a counter-clock-wise orientation.

This is similar to self.project_local_point_and_get_location except that the resulting PointProjection::is_inside will be set to true if the point is inside of the area delimited by this polyline, assuming that:

This polyline isn’t self-crossing.

This polyline is closed with self.indices[i][1] == self.indices[(i + 1) % num_indices][0] where num_indices == self.indices.len().

This polyline is oriented counter-clockwise.

In 3D, the polyline is assumed to be fully coplanar, on a plane with normal given by These properties are not checked.