Trait rapier3d::geometry::SupportMap

pub trait SupportMap {
    fn local_support_point(
        &self, 
        dir: &Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>
    ) -> OPoint<f32, Const<3_usize>>;

    fn local_support_point_toward(
        &self, 
        dir: &Unit<Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>>
    ) -> OPoint<f32, Const<3_usize>> { ... }
    fn support_point(
        &self, 
        transform: &Isometry<f32, Unit<Quaternion<f32>>, 3_usize>, 
        dir: &Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>
    ) -> OPoint<f32, Const<3_usize>> { ... }
    fn support_point_toward(
        &self, 
        transform: &Isometry<f32, Unit<Quaternion<f32>>, 3_usize>, 
        dir: &Unit<Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>>
    ) -> OPoint<f32, Const<3_usize>> { ... }
}

Traits of convex shapes representable by a support mapping function.

Parameters:

V - type of the support mapping direction argument and of the returned point.

Required Methods

fn local_support_point(
&self,
dir: &Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>
) -> OPoint<f32, Const<3_usize>>

Provided Methods

fn local_support_point_toward(
&self,
dir: &Unit<Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>>
) -> OPoint<f32, Const<3_usize>>

Same as self.local_support_point except that dir is normalized.

fn support_point(
&self,
transform: &Isometry<f32, Unit<Quaternion<f32>>, 3_usize>,
dir: &Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>
) -> OPoint<f32, Const<3_usize>>

fn support_point_toward(
&self,
transform: &Isometry<f32, Unit<Quaternion<f32>>, 3_usize>,
dir: &Unit<Matrix<f32, Const<3_usize>, Const<1_usize>, ArrayStorage<f32, 3_usize, 1_usize>>>
) -> OPoint<f32, Const<3_usize>>

Same as self.support_point except that dir is normalized.