Type Alias rapier3d::geometry::RayIntersection

pub type RayIntersection = RayIntersection;

The intersection between a ray and a collider.

Aliased Type

struct RayIntersection {
    pub toi: f32,
    pub normal: Matrix<f32, Const<3>, Const<1>, ArrayStorage<f32, 3, 1>>,
    pub feature: FeatureId,
}

Fields
toi: f32

The time of impact of the ray with the object. The exact contact point can be computed with: ray.point_at(toi) or equivalently origin + dir * toi where origin is the origin of the ray; dir is its direction and toi is the value of this field.

normal: Matrix<f32, Const<3>, Const<1>, ArrayStorage<f32, 3, 1>>

The normal at the intersection point.

If the origin of the ray is inside of the shape and the shape is not solid, the normal will point towards the interior of the shape. Otherwise, the normal points outward.

If the toi is exactly zero, the normal might not be reliable.

feature: FeatureId

Feature at the intersection point.