Module rapier3d::math

Elementary mathematical entities (vectors, matrices, isometries, etc).

Constants

ANG_DIM

The maximum number of rotational degrees of freedom of a rigid-body.

DEFAULT_EPSILON

The default tolerance used for geometric operations.

DIM

The dimension of the space.

MAX_MANIFOLD_POINTS

Max number of pairs of contact points from the same contact manifold that can be solved as part of a single contact constraint.

SIMD_LAST_INDEX

SIMD_WIDTH - 1

SIMD_WIDTH

The number of lanes of a SIMD number.

SPATIAL_DIM

The maximum number of possible rotations and translations of a rigid body.

TWO_DIM

The dimension of the space multiplied by two.

Type Aliases

AngDim

The dimension of the rotations.

AngVector

The angular vector type.

AngularInertia

The angular inertia of a rigid body.

CrossMatrix

A matrix that represent the cross product with a given vector.

Dim

The dimension of the ambient space.

Isometry

The transformation matrix type.

Jacobian

The type of a constraint Jacobian in twist coordinates.

JacobianView

The type of a slice of the constraint Jacobian in twist coordinates.

JacobianViewMut

The type of a mutable slice of the constraint Jacobian in twist coordinates.

Matrix

The matrix type.

Orientation

The orientation type.

Point

The point type.

PrincipalAngularInertia

The principal angular inertia of a rigid body.

Real

The scalar type used throughout this crate.

Rotation

The rotation matrix type.

SdpMatrix

A 3D symmetric-definite-positive matrix.

SimdBool

A SIMD bool with SIMD_WIDTH lanes.

SimdReal

A SIMD float with SIMD_WIDTH lanes.

SpacialVector

A vector with a dimension equal to the maximum number of degrees of freedom of a rigid body.

SpatialDim

The dimension of a spatial vector.

SpatialVector

The vector type with dimension SpatialDim × 1.

Translation

The translation type.

UnitVector

The unit vector type.

Vector

The vector type.