Struct rapier3d::pipeline::QueryFilter

pub struct QueryFilter<'a> {
    pub flags: QueryFilterFlags,
    pub groups: Option<InteractionGroups>,
    pub exclude_collider: Option<ColliderHandle>,
    pub exclude_rigid_body: Option<RigidBodyHandle>,
    pub predicate: Option<&'a dyn Fn(ColliderHandle, &Collider) -> bool>,
}

A filter tha describes what collider should be included or excluded from a scene query.

Fields
flags: QueryFilterFlags

Flags indicating what particular type of colliders should be excluded from the scene query.

groups: Option<InteractionGroups>

If set, only colliders with collision groups compatible with this one will be included in the scene query.

exclude_collider: Option<ColliderHandle>

If set, this collider will be excluded from the scene query.

exclude_rigid_body: Option<RigidBodyHandle>

If set, any collider attached to this rigid-body will be excluded from the scene query.

predicate: Option<&'a dyn Fn(ColliderHandle, &Collider) -> bool>

If set, any collider for which this closure returns false will be excluded from the scene query.

Implementations

impl<'a> QueryFilter<'a>

pub fn test( &self, bodies: &RigidBodySet, handle: ColliderHandle, collider: &Collider ) -> bool

Applies the filters described by self to a collider to determine if it has to be included in a scene query (true) or not (false).

impl<'a> QueryFilter<'a>

pub fn new() -> Self

A query filter that doesn’t exclude any collider.

pub fn exclude_fixed() -> Self

Exclude from the query any collider attached to a fixed rigid-body and colliders with no rigid-body attached.

pub fn exclude_kinematic() -> Self

Exclude from the query any collider attached to a kinematic rigid-body.

pub fn exclude_dynamic() -> Self

Exclude from the query any collider attached to a dynamic rigid-body.

pub fn only_dynamic() -> Self

Excludes all colliders not attached to a dynamic rigid-body.

pub fn only_kinematic() -> Self

Excludes all colliders not attached to a kinematic rigid-body.

pub fn only_fixed() -> Self

Exclude all colliders attached to a non-fixed rigid-body (this will not exclude colliders not attached to any rigid-body).

pub fn exclude_sensors(self) -> Self

Exclude from the query any collider that is a sensor.

pub fn exclude_solids(self) -> Self

Exclude from the query any collider that is not a sensor.

pub fn groups(self, groups: InteractionGroups) -> Self

Only colliders with collision groups compatible with this one will be included in the scene query.

pub fn exclude_collider(self, collider: ColliderHandle) -> Self

Set the collider that will be excluded from the scene query.

pub fn exclude_rigid_body(self, rigid_body: RigidBodyHandle) -> Self

Set the rigid-body that will be excluded from the scene query.

pub fn predicate( self, predicate: &'a impl Fn(ColliderHandle, &Collider) -> bool ) -> Self

Set the predicate to apply a custom collider filtering during the scene query.