Struct rapier3d::pipeline::QueryPipeline

pub struct QueryPipeline { /* private fields */ }

A pipeline for performing queries on all the colliders of a scene.

Implementations

impl QueryPipeline

pub fn new() -> Self

Initializes an empty query pipeline.

pub fn with_query_dispatcher<D>(d: D) -> Selfwhere D: 'static + QueryDispatcher,

Initializes an empty query pipeline with a custom QueryDispatcher.

Use this constructor in order to use a custom QueryDispatcher that is aware of your own user-defined shapes.

pub fn query_dispatcher(&self) -> &dyn QueryDispatcher

The query dispatcher used by this query pipeline for running scene queries.

pub fn update_incremental( &mut self, colliders: &ColliderSet, modified_colliders: &[ColliderHandle], removed_colliders: &[ColliderHandle], refit_and_rebalance: bool )

Update the query pipeline incrementally, avoiding a complete rebuild of its internal data-structure.

pub fn update(&mut self, bodies: &RigidBodySet, colliders: &ColliderSet)

Update the acceleration structure on the query pipeline.

pub fn update_with_mode( &mut self, bodies: &RigidBodySet, colliders: &ColliderSet, mode: QueryPipelineMode )

Update the acceleration structure on the query pipeline.

pub fn cast_ray( &self, bodies: &RigidBodySet, colliders: &ColliderSet, ray: &Ray, max_toi: Real, solid: bool, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, Real)>

Find the closest intersection between a ray and a set of collider.

Parameters

colliders - The set of colliders taking part in this pipeline.

ray: the ray to cast.

max_toi: the maximum time-of-impact that can be reported by this cast. This effectively limits the length of the ray to ray.dir.norm() * max_toi. Use Real::MAX for an unbounded ray.

solid: if this is true an impact at time 0.0 (i.e. at the ray origin) is returned if it starts inside of a shape. If this false then the ray will hit the shape’s boundary even if its starts inside of it.

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn cast_ray_and_get_normal( &self, bodies: &RigidBodySet, colliders: &ColliderSet, ray: &Ray, max_toi: Real, solid: bool, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, RayIntersection)>

Find the closest intersection between a ray and a set of collider.

Parameters

colliders - The set of colliders taking part in this pipeline.

ray: the ray to cast.

max_toi: the maximum time-of-impact that can be reported by this cast. This effectively limits the length of the ray to ray.dir.norm() * max_toi. Use Real::MAX for an unbounded ray.

solid: if this is true an impact at time 0.0 (i.e. at the ray origin) is returned if it starts inside of a shape. If this false then the ray will hit the shape’s boundary even if its starts inside of it.

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn intersections_with_ray<'a>( &self, bodies: &'a RigidBodySet, colliders: &'a ColliderSet, ray: &Ray, max_toi: Real, solid: bool, filter: QueryFilter<'_>, callback: impl FnMut(ColliderHandle, RayIntersection) -> bool )

Find the all intersections between a ray and a set of collider and passes them to a callback.

Parameters

colliders - The set of colliders taking part in this pipeline.

ray: the ray to cast.

max_toi: the maximum time-of-impact that can be reported by this cast. This effectively limits the length of the ray to ray.dir.norm() * max_toi. Use Real::MAX for an unbounded ray.

solid: if this is true an impact at time 0.0 (i.e. at the ray origin) is returned if it starts inside of a shape. If this false then the ray will hit the shape’s boundary even if its starts inside of it.

filter: set of rules used to determine which collider is taken into account by this scene query.

callback: function executed on each collider for which a ray intersection has been found. There is no guarantees on the order the results will be yielded. If this callback returns false, this method will exit early, ignore any further raycast.

pub fn intersection_with_shape( &self, bodies: &RigidBodySet, colliders: &ColliderSet, shape_pos: &Isometry<Real>, shape: &dyn Shape, filter: QueryFilter<'_> ) -> Option<ColliderHandle>

Gets the handle of up to one collider intersecting the given shape.

Parameters

colliders - The set of colliders taking part in this pipeline.

shape_pos - The position of the shape used for the intersection test.

shape - The shape used for the intersection test.

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn project_point( &self, bodies: &RigidBodySet, colliders: &ColliderSet, point: &Point<Real>, solid: bool, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, PointProjection)>

Find the projection of a point on the closest collider.

Parameters

colliders - The set of colliders taking part in this pipeline.

point - The point to project.

solid - If this is set to true then the collider shapes are considered to be plain (if the point is located inside of a plain shape, its projection is the point itself). If it is set to false the collider shapes are considered to be hollow (if the point is located inside of an hollow shape, it is projected on the shape’s boundary).

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn intersections_with_point( &self, bodies: &RigidBodySet, colliders: &ColliderSet, point: &Point<Real>, filter: QueryFilter<'_>, callback: impl FnMut(ColliderHandle) -> bool )

Find all the colliders containing the given point.

Parameters

colliders - The set of colliders taking part in this pipeline.

point - The point used for the containment test.

filter: set of rules used to determine which collider is taken into account by this scene query.

callback - A function called with each collider with a shape containing the point.

pub fn project_point_and_get_feature( &self, bodies: &RigidBodySet, colliders: &ColliderSet, point: &Point<Real>, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, PointProjection, FeatureId)>

Find the projection of a point on the closest collider.

The results include the ID of the feature hit by the point.

Parameters

colliders - The set of colliders taking part in this pipeline.

point - The point to project.

solid - If this is set to true then the collider shapes are considered to be plain (if the point is located inside of a plain shape, its projection is the point itself). If it is set to false the collider shapes are considered to be hollow (if the point is located inside of an hollow shape, it is projected on the shape’s boundary).

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn colliders_with_aabb_intersecting_aabb( &self, aabb: &Aabb, callback: impl FnMut(&ColliderHandle) -> bool )

Finds all handles of all the colliders with an Aabb intersecting the given Aabb.

pub fn cast_shape<'a>( &self, bodies: &RigidBodySet, colliders: &'a ColliderSet, shape_pos: &Isometry<Real>, shape_vel: &Vector<Real>, shape: &dyn Shape, max_toi: Real, stop_at_penetration: bool, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, TOI)>

Casts a shape at a constant linear velocity and retrieve the first collider it hits.

This is similar to ray-casting except that we are casting a whole shape instead of just a point (the ray origin). In the resulting TOI, witness and normal 1 refer to the world collider, and are in world space.

Parameters

colliders - The set of colliders taking part in this pipeline.

shape_pos - The initial position of the shape to cast.

shape_vel - The constant velocity of the shape to cast (i.e. the cast direction).

shape - The shape to cast.

max_toi - The maximum time-of-impact that can be reported by this cast. This effectively limits the distance traveled by the shape to shapeVel.norm() * maxToi.

stop_at_penetration - If set to false, the linear shape-cast won’t immediately stop if the shape is penetrating another shape at its starting point and its trajectory is such that it’s on a path to exist that penetration state.

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn nonlinear_cast_shape( &self, bodies: &RigidBodySet, colliders: &ColliderSet, shape_motion: &NonlinearRigidMotion, shape: &dyn Shape, start_time: Real, end_time: Real, stop_at_penetration: bool, filter: QueryFilter<'_> ) -> Option<(ColliderHandle, TOI)>

Casts a shape with an arbitrary continuous motion and retrieve the first collider it hits.

In the resulting TOI, witness and normal 1 refer to the world collider, and are in world space.

Parameters

colliders - The set of colliders taking part in this pipeline.

shape_motion - The motion of the shape.

shape - The shape to cast.

start_time - The starting time of the interval where the motion takes place.

end_time - The end time of the interval where the motion takes place.

stop_at_penetration - If the casted shape starts in a penetration state with any collider, two results are possible. If stop_at_penetration is true then, the result will have a toi equal to start_time. If stop_at_penetration is false then the nonlinear shape-casting will see if further motion wrt. the penetration normal would result in tunnelling. If it does not (i.e. we have a separating velocity along that normal) then the nonlinear shape-casting will attempt to find another impact, at a time > start_time that could result in tunnelling.

filter: set of rules used to determine which collider is taken into account by this scene query.

pub fn intersections_with_shape<'a>( &self, bodies: &RigidBodySet, colliders: &'a ColliderSet, shape_pos: &Isometry<Real>, shape: &dyn Shape, filter: QueryFilter<'_>, callback: impl FnMut(ColliderHandle) -> bool )

Retrieve all the colliders intersecting the given shape.

Parameters

colliders - The set of colliders taking part in this pipeline.

shapePos - The position of the shape to test.

shapeRot - The orientation of the shape to test.

shape - The shape to test.

filter: set of rules used to determine which collider is taken into account by this scene query.

callback - A function called with the handles of each collider intersecting the shape.