Module rapier3d::utils

Miscellaneous utilities.

Traits

IndexMut2

Methods for simultaneously indexing a container with two distinct indices.

WBasis

Trait to compute the orthonormal basis of a vector.

WQuat

Trait implemented by quaternions.

WReal

The trait for real numbers used by Rapier.

WSign

Trait to copy the sign of each component of one scalar/vector/matrix to another.