Trait rapier3d::utils::WBasis

pub trait WBasis: Sized {
    type Basis;

    fn orthonormal_basis(self) -> Self::Basis;
    fn orthonormal_vector(self) -> Self;
}

Trait to compute the orthonormal basis of a vector.

Required Associated Types

type Basis

The type of the array of orthonormal vectors.

Required Methods

fn orthonormal_basis(self) -> Self::Basis

Computes the vectors which, when combined with self, form an orthonormal basis.

fn orthonormal_vector(self) -> Self

Computes a vector orthogonal to self with a unit length (if self has a unit length).