Trait rapier3d::utils::WSign

pub trait WSign<Rhs>: Sized {
    // Required method
    fn copy_sign_to(self, to: Rhs) -> Rhs;
}

Trait to copy the sign of each component of one scalar/vector/matrix to another.

Required Methods

fn copy_sign_to(self, to: Rhs) -> Rhs

Copy the sign of each component of self to the corresponding component of to.

Object Safety

This trait is not object safe.