Struct rapier3d::dynamics::IntegrationParameters

pub struct IntegrationParameters {

    pub dt: Real,
    pub min_ccd_dt: Real,
    pub erp: Real,
    pub damping_ratio: Real,
    pub joint_erp: Real,
    pub joint_damping_ratio: Real,
    pub allowed_linear_error: Real,
    pub max_penetration_correction: Real,
    pub prediction_distance: Real,
    pub max_velocity_iterations: usize,
    pub max_velocity_friction_iterations: usize,
    pub max_stabilization_iterations: usize,
    pub interleave_restitution_and_friction_resolution: bool,
    pub min_island_size: usize,
    pub max_ccd_substeps: usize,

}

Parameters for a time-step of the physics engine.

Fields
dt: Real

The timestep length (default: 1.0 / 60.0)

min_ccd_dt: Real

Minimum timestep size when using CCD with multiple substeps (default 1.0 / 60.0 / 100.0)

When CCD with multiple substeps is enabled, the timestep is subdivided into smaller pieces. This timestep subdivision wonโt generate timestep lengths smaller than min_ccd_dt.

Setting this to a large value will reduce the opportunity to performing CCD substepping, resulting in potentially more time dropped by the motion-clamping mechanism. Setting this to an very small value may lead to numerical instabilities.

erp: Real

0-1: multiplier for how much of the constraint violation (e.g. contact penetration) will be compensated for during the velocity solve. (default 0.8).

damping_ratio: Real

0-1: the damping ratio used by the springs for Baumgarte constraints stabilization. Lower values make the constraints more compliant (more โspringyโ, allowing more visible penetrations before stabilization). (default 0.25).

joint_erp: Real

0-1: multiplier for how much of the joint violation will be compensated for during the velocity solve. (default 1.0).

joint_damping_ratio: Real

The fraction of critical damping applied to the joint for constraints regularization. (default 0.25).

allowed_linear_error: Real

Amount of penetration the engine wont attempt to correct (default: 0.001m).

max_penetration_correction: Real

Maximum amount of penetration the solver will attempt to resolve in one timestep.

prediction_distance: Real

The maximal distance separating two objects that will generate predictive contacts (default: 0.002).

max_velocity_iterations: usize

Maximum number of iterations performed to solve non-penetration and joint constraints (default: 4).

max_velocity_friction_iterations: usize

Maximum number of iterations performed to solve friction constraints (default: 8).

max_stabilization_iterations: usize

Maximum number of iterations performed to remove the energy introduced by penetration corrections (default: 1).

interleave_restitution_and_friction_resolution: bool

If false, friction and non-penetration constraints will be solved in the same loop. Otherwise, non-penetration constraints are solved first, and friction constraints are solved after (default: true).

min_island_size: usize

Minimum number of dynamic bodies in each active island (default: 128).

max_ccd_substeps: usize

Maximum number of substeps performed by the solver (default: 1).

Implementations

impl IntegrationParameters

pub fn inv_dt(&self) -> Real

The inverse of the time-stepping length, i.e. the steps per seconds (Hz).

This is zero if self.dt is zero.

pub fn set_dt(&mut self, dt: Real)

๐Deprecated: You can just set the IntegrationParams::dt value directly

Sets the time-stepping length.

pub fn set_inv_dt(&mut self, inv_dt: Real)

Sets the inverse time-stepping length (i.e. the frequency).

This automatically recompute self.dt.

pub fn erp_inv_dt(&self) -> Real

The ERP coefficient, multiplied by the inverse timestep length.

pub fn joint_erp_inv_dt(&self) -> Real

The joint ERP coefficient, multiplied by the inverse timestep length.

pub fn cfm_factor(&self) -> Real

The CFM factor to be used in the constraints resolution.

pub fn joint_cfm_coeff(&self) -> Real

The CFM (constraints force mixing) coefficient applied to all joints for constraints regularization