Struct rapier3d::geometry::SharedShape

pub struct SharedShape(pub Arc<dyn Shape, Global>);

The shape of a collider.

Tuple Fields
0: Arc<dyn Shape, Global>
Implementations

impl SharedShape

pub fn new(shape: impl Shape) -> SharedShape

Wraps the given shape as a shared shape.

pub fn make_mut(&mut self) -> &mut (dyn Shape + 'static)

If this shape is shared, then the content of self is cloned into a unique instance, and a mutable reference to that instance is returned.

pub fn compound( shapes: Vec<(Isometry<f32, Unit<Quaternion<f32>>, 3>, SharedShape), Global> ) -> SharedShape

Initialize a compound shape defined by its subshapes.

pub fn ball(radius: f32) -> SharedShape

Initialize a ball shape defined by its radius.

pub fn halfspace( outward_normal: Unit<Matrix<f32, Const<3>, Const<1>, ArrayStorage<f32, 3, 1>>> ) -> SharedShape

Initialize a plane shape defined by its outward normal.

pub fn cylinder(half_height: f32, radius: f32) -> SharedShape

Initialize a cylindrical shape defined by its half-height (along along the y axis) and its radius.

pub fn round_cylinder( half_height: f32, radius: f32, border_radius: f32 ) -> SharedShape

Initialize a rounded cylindrical shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn round_cone( half_height: f32, radius: f32, border_radius: f32 ) -> SharedShape

Initialize a rounded cone shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn cone(half_height: f32, radius: f32) -> SharedShape

Initialize a cone shape defined by its half-height (along along the y axis) and its basis radius.

pub fn cuboid(hx: f32, hy: f32, hz: f32) -> SharedShape

Initialize a cuboid shape defined by its half-extents.

pub fn round_cuboid( hx: f32, hy: f32, hz: f32, border_radius: f32 ) -> SharedShape

Initialize a round cuboid shape defined by its half-extents and border radius.

pub fn capsule( a: OPoint<f32, Const<3>>, b: OPoint<f32, Const<3>>, radius: f32 ) -> SharedShape

Initialize a capsule shape from its endpoints and radius.

pub fn capsule_x(half_height: f32, radius: f32) -> SharedShape

Initialize a capsule shape aligned with the x axis.

pub fn capsule_y(half_height: f32, radius: f32) -> SharedShape

Initialize a capsule shape aligned with the y axis.

pub fn capsule_z(half_height: f32, radius: f32) -> SharedShape

Initialize a capsule shape aligned with the z axis.

pub fn segment( a: OPoint<f32, Const<3>>, b: OPoint<f32, Const<3>> ) -> SharedShape

Initialize a segment shape from its endpoints.

pub fn triangle( a: OPoint<f32, Const<3>>, b: OPoint<f32, Const<3>>, c: OPoint<f32, Const<3>> ) -> SharedShape

Initializes a triangle shape.

pub fn round_triangle( a: OPoint<f32, Const<3>>, b: OPoint<f32, Const<3>>, c: OPoint<f32, Const<3>>, border_radius: f32 ) -> SharedShape

Initializes a triangle shape with round corners.

pub fn polyline( vertices: Vec<OPoint<f32, Const<3>>, Global>, indices: Option<Vec<[u32; 2], Global>> ) -> SharedShape

Initializes a polyline shape defined by its vertex and index buffers.

If no index buffer is provided, the polyline is assumed to describe a line strip.

pub fn trimesh( vertices: Vec<OPoint<f32, Const<3>>, Global>, indices: Vec<[u32; 3], Global> ) -> SharedShape

Initializes a triangle mesh shape defined by its vertex and index buffers.

pub fn trimesh_with_flags( vertices: Vec<OPoint<f32, Const<3>>, Global>, indices: Vec<[u32; 3], Global>, flags: TriMeshFlags ) -> SharedShape

Initializes a triangle mesh shape defined by its vertex and index buffers and pre-processing flags.

pub fn convex_decomposition( vertices: &[OPoint<f32, Const<3>>], indices: &[[u32; 3]] ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition( vertices: &[OPoint<f32, Const<3>>], indices: &[[u32; 3]], border_radius: f32 ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_decomposition_with_params( vertices: &[OPoint<f32, Const<3>>], indices: &[[u32; 3]], params: &VHACDParameters ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition_with_params( vertices: &[OPoint<f32, Const<3>>], indices: &[[u32; 3]], params: &VHACDParameters, border_radius: f32 ) -> SharedShape

Initializes a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_hull(points: &[OPoint<f32, Const<3>>]) -> Option<SharedShape>

Creates a new shared shape that is the convex-hull of the given points.

pub fn convex_mesh( points: Vec<OPoint<f32, Const<3>>, Global>, indices: &[[u32; 3]] ) -> Option<SharedShape>

Creates a new shared shape that is a convex polyhedron formed by the given set of points assumed to form a convex mesh (no convex-hull will be automatically computed).

pub fn round_convex_hull( points: &[OPoint<f32, Const<3>>], border_radius: f32 ) -> Option<SharedShape>

Creates a new shared shape with rounded corners that is the convex-hull of the given points, dilated by border_radius.

pub fn round_convex_mesh( points: Vec<OPoint<f32, Const<3>>, Global>, indices: &[[u32; 3]], border_radius: f32 ) -> Option<SharedShape>

Creates a new shared shape with round corners that is a convex polyhedron formed by the given set of points assumed to form a convex mesh (no convex-hull will be automatically computed).

pub fn heightfield( heights: Matrix<f32, Dyn, Dyn, VecStorage<f32, Dyn, Dyn>>, scale: Matrix<f32, Const<3>, Const<1>, ArrayStorage<f32, 3, 1>> ) -> SharedShape

Initializes an heightfield shape on the x-z plane defined by its set of height and a scale factor along each coordinate axis.