Trait rapier3d::geometry::SimdCompositeShape

pub trait SimdCompositeShape {
    // Required methods
    fn map_part_at(
        &self,
        shape_id: u32,
        f: &mut dyn FnMut(Option<&Isometry<f32, Unit<Quaternion<f32>>, 3>>, &(dyn Shape + 'static))
    );
    fn qbvh(&self) -> &GenericQbvh<u32, DefaultStorage>;
}

Trait implemented by shapes composed of multiple simpler shapes.

A composite shape is composed of several shapes. For example, this can be a convex decomposition of a concave shape; or a triangle-mesh.

Required Methods

fn map_part_at( &self, shape_id: u32, f: &mut dyn FnMut(Option<&Isometry<f32, Unit<Quaternion<f32>>, 3>>, &(dyn Shape + 'static)) )

Applies a function to one sub-shape of this composite shape.

fn qbvh(&self) -> &GenericQbvh<u32, DefaultStorage>

Gets the acceleration structure of the composite shape.